Use inheritance for all DashCounters
