+++ orchagent/dash/dashcounter.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <map>
-#include <mutex>
 #include <string>
 #include <unordered_set>
 
@@ -10,6 +9,7 @@
 #include "flex_counter_manager.h"
 
 #include "dash_api/eni.pb.h"
+#include "dash_api/ha_set.pb.h"
 
 struct EniEntry
 {
@@ -25,7 +25,6 @@ struct DashCounter
     FlexCounterManager stat_manager;
     bool fc_status = false;
     std::unordered_set<std::string> counter_stats;
-    std::once_flag fetch_stats_flag;
 
     DashCounter() {}
     DashCounter(const std::string& group_name, StatsMode stats_mode, uint polling_interval, bool enabled) 
@@ -56,6 +55,12 @@ struct DashCounter
         }
         stat_manager.clearCounterIdList(oid);
     }
+};
+
+struct DashEniCounter: DashCounter<CounterType::ENI>
+{
+    DashEniCounter(const std::string& group_name, StatsMode stats_mode, uint polling_interval, bool enabled) 
+        : DashCounter<CounterType::ENI>(group_name, stats_mode, polling_interval, enabled) {}
 
     void refreshStats(bool install, const EniTable& eni_entries)
     {
@@ -79,6 +84,76 @@ struct DashCounter
         if (fc_status != prev_enabled)
         {
             refreshStats(fc_status, eni_entries);
+        }
+    }
+};
+
+struct DashMeterCounter: DashCounter<CounterType::DASH_METER>
+{
+    DashMeterCounter(const std::string& group_name, StatsMode stats_mode, uint polling_interval, bool enabled) 
+        : DashCounter<CounterType::DASH_METER>(group_name, stats_mode, polling_interval, enabled) {}
+    
+    void refreshStats(bool install, const EniTable& eni_entries)
+    {
+        for (auto it = eni_entries.begin(); it != eni_entries.end(); it++)
+        {
+            if (install)
+            {
+                addToFC(it->second.eni_id, it->first);
+            }
+            else
+            {
+                removeFromFC(it->second.eni_id, it->first);
+            }
+        }
+    }
+
+    void handleStatusUpdate(bool enabled, const EniTable& eni_entries)
+    {
+        bool prev_enabled = fc_status;
+        fc_status = enabled;
+        if (fc_status != prev_enabled)
+        {
+            refreshStats(fc_status, eni_entries);
+        }
+    }
+    };
+
+struct HaSetEntry
+{
+    sai_object_id_t ha_set_id;
+    dash::ha_set::HaSet metadata;
+};
+
+typedef std::map<std::string, HaSetEntry> HaSetTable;
+
+struct DashHaCounter: DashCounter<CounterType::HA_SET>
+{
+    DashHaCounter(const std::string& group_name, StatsMode stats_mode, uint polling_interval, bool enabled) 
+        : DashCounter<CounterType::HA_SET>(group_name, stats_mode, polling_interval, enabled) {}
+    
+    void refreshStats(bool install, const HaSetTable& ha_set_entries)
+    {
+        for (auto it = ha_set_entries.begin(); it != ha_set_entries.end(); it++)
+        {
+            if (install)
+            {
+                addToFC(it->second.ha_set_id, it->first);
+            }
+            else
+            {
+                removeFromFC(it->second.ha_set_id, it->first);
+            }
+        }
+    }
+
+    void handleStatusUpdate(bool enabled, const HaSetTable& ha_set_entries)
+    {
+        bool prev_enabled = fc_status;
+        fc_status = enabled;
+        if (fc_status != prev_enabled)
+        {
+            refreshStats(fc_status, ha_set_entries);
         }
     }
 };
